backend/ai_requirements.txt
flask
flask-cors
google-generativeai
pillow
requests
python-dotenv